+++ libs/estdlib/src/erlang.erl
@@ -38,7 +38,8 @@
     map_get/2,
     map_is_key/2,
     min/2,
-    max/2
+    max/2,
+    memory/1
 ]).
 
 %%
@@ -47,6 +48,8 @@
 %% * return value needs to be {ok, cancel} or {error, Reason}
 %% * review API documentation for timer functions in this module
 %%
+
+-type mem_type() :: binary.
 
 %%-----------------------------------------------------------------------------
 %% @param   Time time in milliseconds after which to send the timeout message.
@@ -281,3 +284,14 @@ min(_A, B) -> B.
 %%-----------------------------------------------------------------------------
 max(A, B) when A > B -> A;
 max(_A, B) -> B.
+
+%%-----------------------------------------------------------------------------
+%% @param   Type the type of memory to request
+%% @returns the amount of memory (in bytes) used of the specified type
+%% @doc     Return the amount of memory (in bytes) used of the specified type
+%%
+%% @end
+%%-----------------------------------------------------------------------------
+-spec memory(Type :: mem_type()) -> non_neg_integer().
+memory(_Type) ->
+    throw(nif_error).